Network: Move testing/evaluation utilities into separate package
Also refactor existing scripts using them.
